requirements.txt
streamlit
speechrecognition
pyaudio
gtts
groq
langchain-groq
python-dotenv